pyproject.toml
[tool.poetry]
name = "slack-bot"
version = "0.1.0"
description = "Slack Bot"
authors = ["Aditya Ghaffar <adit.bisa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
slack-bolt = "^1.18.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
